Remove dependencies of TUnixSystem & TWinNTSystem on /net/inc
Move declaration of ESockOptions and ESendRecvOptions enums from TSocket.h to TSystem.h, to get rid of this include dependencies in CMake:
include_directories(${CMAKE_SOURCE_DIR}/net/net/inc)
